Python Folder/age_and_gender_classification/requirements.txt
# Core numerical
numpy==1.23.5
# Computer Vision
opencv-python==4.10.0.84

#SQL Connection
pymysql==1.1.2

# Deep Learning
tensorflow==2.10.0
torch==2.4.1
torchvision==0.19.1
timm==1.0.12

# Models & Utils
pillow==10.4.0
protobuf==3.19.6

# Optional (only if used)
mediapipe==0.10.8
